sadsads

+++ python_variable_annotations/2-floor.py
@@ -8,4 +8,4 @@ def floor(floor: float) -> int:
     """
     sadasdsa
     """
-    return floor
+    return int(floor)
